test: add typescript e2e test without babel
Adds a test close to the setup used by the CLI in a TS project without Babel.

The existing TS test was moved to `typescript-with-babel`.
The new test is a copy of the older one, with all babel dependencies removed, a rewritten component using the Composition API, a test in TS, a TS config close to the CLI one, and a shim file.