feat: use TS config to have commonjs modules
The previous logic was trying to figure out if CommonJS modules were used in the TS config and if not,
and if no babel configuration provided, then loaded a babel transformer to change the module format.

This changes the logic to use a more straightforward method by directly updating the tsconfig to commonjs,
which should hopefully have the same results without side-effects.